Add a typescript rule so that array<T>[i] is T | undefined
Because i could be out of bounds (or even: not an integer).

This rule is sometimes really annoying (some call it pedantic), but I
think it can help find bugs.

I'm enabling it as a distinct PR, before a bigger PR where I upgrade
hightable, and need to add code to sort the rows along multiple columns,
where I prefer to have this rule enabled, as in hightable.